refactor navbar components
