+++ src/app.jsx
@@ -33,15 +33,11 @@ import {
   PurchaseDetail,
   Wishlist
 } from './pages/user'
-
 import { CartFloatButton, DarkModeButton } from './components/buttons'
-
-import Navbar from './components/navbar'
+import { NavDrawer, Navbar } from './components/navbars'
 import Footer from './components/footer'
-import NavDrawer from './components/navDrawer'
 import NavMode from './hooks/useNavMode'
 import Sidebar from './components/user/sidebar'
-
 import SearchMode from './hooks/useSearchMode'
 import SearchDrawer from './components/search/searchDrawer'
 import { auth } from './services/firebase.config'
