+++ src/components/navbars/navDrawer.jsx
@@ -1,14 +1,14 @@
-import { useDarkMode } from '../hooks/useDarkMode'
-import { useNavMode } from '../hooks/useNavMode'
-import EmailField from './emailField'
-import LinkButton from './buttons/linkButton'
-import SocialMedia from './socialMedia'
+import { useDarkMode } from '../../hooks/useDarkMode'
+import { useNavMode } from '../../hooks/useNavMode'
+import EmailField from '../emailField'
+import LinkButton from '../buttons/linkButton'
+import SocialMedia from '../socialMedia'
 import { useDispatch, useSelector } from 'react-redux'
 import {
   selectCategoryStatus,
   selectCategories,
   getCategories
-} from '../redux/features/category/categorySlice'
+} from '../../redux/features/category/categorySlice'
 import { useEffect } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faToggleOff, faToggleOn } from '@fortawesome/free-solid-svg-icons'
